libpoll 1.00.01

+++ source/includes/libpoll-wrapper.h
@@ -64,7 +64,7 @@ typedef struct _stpolConnectInfo {
 		@param connect2ndbuffer	This is a dynamic buffer storage that will expand as required by the data allocation size needed to be sent. 
 		@return					pol base.
 */
-polbase * polnewbase(polloghandler loghandler=0, DWORD logverboseflags=-1, int connect2ndbuffer=0);
+polbase * polnewbase(polloghandler loghandler=0, DWORD logverboseflags=(DWORD)-1, int connect2ndbuffer=0);
 
 /**
 	Start accepting connections.
